refactor: Move dashboard components into a folder
