+++ uptick-dashboard/app/components/dashboard/navbar.js
