+++ uptick-dashboard/app/components/dashboard/sidenav.js
